requirements.txt
python-dotenv>=1.0.0
pymongo>=4.6.0
qdrant-client>=1.7.0
langchain>=0.1.0
fastembed>=0.3.0
firecrawl-py>=0.0.16
streamlit>=1.28.0
google-generativeai>=0.3.0